refactor(validation): move `@sanity/validation` into monopackage
Rationale: the exported API has never really been built for external consumption.
It requires a number of internal things that is very cumbersome to pull together
outside of a Sanity application - schema, client getters, document existence
checks and whatnot.

By moving it to the monopackage, we can more easily treat it as what it is;
an internal API. With time, we can instead work on exposing a simplified API
meant for external consumption in node scripts and similar.
