--- packages/@sanity/validation/.depcheckignore.json
@@ -1,3 +1,0 @@
-{
-  "ignore": ["@optimize-lodash/rollup-plugin", "@sanity/pkg-utils"]
-}
