--- packages/@sanity/validation/.gitignore
@@ -1,12 +1,0 @@
-# Logs
-/logs
-*.log
-
-# Coverage directory used by tools like istanbul
-/coverage
-
-# Dependency directories
-/node_modules
-
-# Compiled code
-/lib
