--- packages/@sanity/validation/jest.config.cjs
@@ -1,7 +1,0 @@
-'use strict'
-
-const {createJestConfig} = require('../../../test/config.cjs')
-
-module.exports = createJestConfig({
-  displayName: require('./package.json').name,
-})
