--- packages/@sanity/validation/package.config.ts
@@ -1,4 +1,0 @@
-import {defineConfig} from '@sanity/pkg-utils'
-import baseConfig from '../../../package.config'
-
-export default defineConfig({...baseConfig})
