--- packages/@sanity/validation/package.json
@@ -1,70 +1,0 @@
-{
-  "name": "@sanity/validation",
-  "version": "3.14.4",
-  "description": "Validation and warning infrastructure for Sanity projects",
-  "keywords": [
-    "sanity",
-    "cms",
-    "headless",
-    "realtime",
-    "content",
-    "validation"
-  ],
-  "homepage": "https://www.sanity.io/",
-  "bugs": {
-    "url": "https://github.com/sanity-io/sanity/issues"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/sanity-io/sanity.git",
-    "directory": "packages/@sanity/validation"
-  },
-  "license": "MIT",
-  "author": "Sanity.io <hello@sanity.io>",
-  "exports": {
-    ".": {
-      "types": "./lib/dts/src/index.d.ts",
-      "source": "./src/index.ts",
-      "require": "./lib/index.js",
-      "node": {
-        "import": "./lib/index.cjs.mjs",
-        "require": "./lib/index.js"
-      },
-      "import": "./lib/index.esm.js",
-      "default": "./lib/index.esm.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "main": "./lib/index.js",
-  "module": "./lib/index.esm.js",
-  "source": "./src/index.ts",
-  "types": "./lib/dts/src/index.d.ts",
-  "files": [
-    "lib",
-    "src"
-  ],
-  "scripts": {
-    "prebuild": "run-s clean",
-    "build": "pkg-utils build --tsconfig tsconfig.lib.json",
-    "postbuild": "run-s check:package",
-    "check:package": "pkg-utils --strict --tsconfig tsconfig.lib.json",
-    "clean": "rimraf lib",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "watch": "pkg-utils watch --tsconfig tsconfig.lib.json"
-  },
-  "dependencies": {
-    "@sanity/types": "3.14.4",
-    "date-fns": "^2.26.1",
-    "lodash": "^4.17.21",
-    "rxjs": "^7.8.0"
-  },
-  "devDependencies": {
-    "@sanity/client": "^6.1.5",
-    "@sanity/schema": "3.14.4"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "a7b7dad6bdfde05883a61f78e998b2571dfbe0e2"
-}
