--- packages/@sanity/validation/test/helpers/nullExport.ts
@@ -1,3 +1,0 @@
-// used to mock out a part
-// see the jest config
-export default null
