--- packages/@sanity/validation/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.settings",
-  "include": ["./src", "./test"],
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": ".",
-    "outDir": "./lib/dts"
-  },
-  "references": [{"path": "../types/tsconfig.lib.json"}]
-}
