--- packages/@sanity/validation/tsconfig.lib.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../../tsconfig.settings",
-  "include": ["./src"],
-  "exclude": [
-    "./src/**/__fixtures__",
-    "./src/**/__mocks__",
-    "./src/**/__workshop__",
-    "./src/**/*.test.ts",
-    "./src/**/*.test.tsx"
-  ],
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": ".",
-    "outDir": "./lib/dts"
-  },
-  "references": [{"path": "../types/tsconfig.lib.json"}]
-}
