+++ packages/sanity/src/core/validation/util/convertToValidationMarker.ts
