+++ packages/sanity/src/core/validation/util/deepEquals.ts
