+++ packages/sanity/src/core/validation/util/escapeRegex.ts
