+++ packages/sanity/src/core/validation/util/pathToString.ts
