+++ packages/sanity/src/core/validation/util/typeString.ts
