+++ packages/sanity/src/core/validation/validateDocument.ts
