+++ packages/sanity/src/core/validation/validators/booleanValidator.ts
