+++ packages/sanity/src/core/validation/validators/objectValidator.ts
