+++ packages/sanity/src/core/validation/validators/slugValidator.ts
