+++ packages/sanity/src/core/validation/validators/stringValidator.ts
